Moved CommonPluginManagerTrait to src/Test, updated docs
